+++ Tests/Command/InvokableCommandTest.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Console\Tests\Command;
 
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Attribute\Argument;
 use Symfony\Component\Console\Attribute\Option;
@@ -18,6 +19,7 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Completion\CompletionInput;
 use Symfony\Component\Console\Completion\CompletionSuggestions;
 use Symfony\Component\Console\Completion\Suggestion;
+use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Symfony\Component\Console\Exception\InvalidOptionException;
 use Symfony\Component\Console\Exception\LogicException;
 use Symfony\Component\Console\Input\ArrayInput;
@@ -130,6 +132,88 @@ class InvokableCommandTest extends TestCase
         self::assertTrue($optInputOption->isValueOptional());
         self::assertFalse($optInputOption->isNegatable());
         self::assertFalse($optInputOption->getDefault());
+    }
+
+    public function testEnumArgument()
+    {
+        $command = new Command('foo');
+        $command->setCode(function (
+            #[Argument] StringEnum $enum,
+            #[Argument] StringEnum $enumWithDefault = StringEnum::Image,
+            #[Argument] ?StringEnum $nullableEnum = null,
+        ): int {
+            Assert::assertSame(StringEnum::Image, $enum);
+            Assert::assertSame(StringEnum::Image, $enumWithDefault);
+            Assert::assertNull($nullableEnum);
+
+            return 0;
+        });
+
+        $enumInputArgument = $command->getDefinition()->getArgument('enum');
+        self::assertTrue($enumInputArgument->isRequired());
+        self::assertNull($enumInputArgument->getDefault());
+        self::assertTrue($enumInputArgument->hasCompletion());
+
+        $enumWithDefaultInputArgument = $command->getDefinition()->getArgument('enum-with-default');
+        self::assertFalse($enumWithDefaultInputArgument->isRequired());
+        self::assertSame('image', $enumWithDefaultInputArgument->getDefault());
+        self::assertTrue($enumWithDefaultInputArgument->hasCompletion());
+
+        $nullableEnumInputArgument = $command->getDefinition()->getArgument('nullable-enum');
+        self::assertFalse($nullableEnumInputArgument->isRequired());
+        self::assertNull($nullableEnumInputArgument->getDefault());
+        self::assertTrue($nullableEnumInputArgument->hasCompletion());
+
+        $enumInputArgument->complete(CompletionInput::fromTokens([], 0), $suggestions = new CompletionSuggestions());
+        self::assertEquals([new Suggestion('image'), new Suggestion('video')], $suggestions->getValueSuggestions());
+
+        $command->run(new ArrayInput(['enum' => 'image']), new NullOutput());
+
+        self::expectException(InvalidArgumentException::class);
+        self::expectExceptionMessage('The value "incorrect" is not valid for the "enum" argument. Supported values are "image", "video".');
+
+        $command->run(new ArrayInput(['enum' => 'incorrect']), new NullOutput());
+    }
+
+    public function testEnumOption()
+    {
+        $command = new Command('foo');
+        $command->setCode(function (
+            #[Option] StringEnum $enum = StringEnum::Video,
+            #[Option] StringEnum $enumWithDefault = StringEnum::Image,
+            #[Option] ?StringEnum $nullableEnum = null,
+        ): int {
+            Assert::assertSame(StringEnum::Image, $enum);
+            Assert::assertSame(StringEnum::Image, $enumWithDefault);
+            Assert::assertNull($nullableEnum);
+
+            return 0;
+        });
+
+        $enumInputOption = $command->getDefinition()->getOption('enum');
+        self::assertTrue($enumInputOption->isValueRequired());
+        self::assertSame('video', $enumInputOption->getDefault());
+        self::assertTrue($enumInputOption->hasCompletion());
+
+        $enumWithDefaultInputOption = $command->getDefinition()->getOption('enum-with-default');
+        self::assertTrue($enumWithDefaultInputOption->isValueRequired());
+        self::assertSame('image', $enumWithDefaultInputOption->getDefault());
+        self::assertTrue($enumWithDefaultInputOption->hasCompletion());
+
+        $nullableEnumInputOption = $command->getDefinition()->getOption('nullable-enum');
+        self::assertTrue($nullableEnumInputOption->isValueRequired());
+        self::assertNull($nullableEnumInputOption->getDefault());
+        self::assertTrue($nullableEnumInputOption->hasCompletion());
+
+        $enumInputOption->complete(CompletionInput::fromTokens([], 0), $suggestions = new CompletionSuggestions());
+        self::assertEquals([new Suggestion('image'), new Suggestion('video')], $suggestions->getValueSuggestions());
+
+        $command->run(new ArrayInput(['--enum' => 'image']), new NullOutput());
+
+        self::expectException(InvalidOptionException::class);
+        self::expectExceptionMessage('The value "incorrect" is not valid for the "enum" option. Supported values are "image", "video".');
+
+        $command->run(new ArrayInput(['--enum' => 'incorrect']), new NullOutput());
     }
 
     public function testInvalidArgumentType()
@@ -376,4 +460,10 @@ class InvokableCommandTest extends TestCase
     {
         return ['ROLE_ADMIN', 'ROLE_USER'];
     }
+}
+
+enum StringEnum: string
+{
+    case Image = 'image';
+    case Video = 'video';
 }